Fix off by one error during selection of whole lines
When a selection includes the EOL character, the `githubinator` command
doesn't compute the `end_line` index correctly. This is because when a
selection includes and ends at an EOL character on line N, the `#rowcol`
method returns line and column numbers `(N+1, 0)`.

Thus, we shouldn't increment `end_line`, since Sublime does this by
itself when a selection ends at a line boundary.